refactor: Trial -> Dto로 변경
